requirements.txt
openai>=1.3.0
PyQt5>=5.15.0
numpy>=1.24.0
sounddevice>=0.4.6
python-dotenv>=1.0.0
pyaudio>=0.2.13
keyboard>=0.13.5
flask>=2.3.3
gunicorn>=21.2.0
werkzeug>=2.3.7
pyopenssl>=23.0.0  # For SSL support
cryptography>=41.0.0  # Required for ad-hoc SSL certificates